+++ components/radio/Radio.vue
@@ -1,34 +1,102 @@
 <script setup lang="ts">
+import { computed, ref, watchEffect } from 'vue'
 interface Option {
   label: string // 选项名
-  value: any // 选项值
-  disabled?: boolean // 是否禁用单个单选器，默认 false
+  value: string | number | boolean // 选项值
+  disabled?: boolean // 是否禁用选项
 }
 interface Props {
   options?: Option[] // 单选框选项数据
   disabled?: boolean // 是否禁用
   vertical?: boolean // 是否垂直排列，仅当 button: false 时生效
-  value?: any // (v-model) 当前选中的值
-  gap?: number // 多个单选框之间的间距，单位px，垂直排列时，间距即垂直间距，仅当 button: false 时生效
+  checked?: boolean // (v-model) 当前是否选中
+  gap?: number | number[] // 多个单选框之间的间距；垂直排列时为垂直间距，单位 px；数组间距用于水平排列折行时：[水平间距, 垂直间距]；仅当 button: false 时生效
+  width?: string | number // 单选区域最大宽度，超出后折行，单位 px；仅当 button: false 时生效
+  height?: string | number // 单选区域最大高度，超出后滚动，单位 px；仅当 button: false 时生效
   button?: boolean // 是否启用按钮样式
   buttonStyle?: 'outline' | 'solid' // 按钮样式风格
-  buttonSize?: 'small' | 'middle' | 'large' // 按钮大小，仅当 button: true 时生效
+  buttonSize?: 'small' | 'middle' | 'large' // 按钮大小；仅当 button: true 时生效
+  value?: string | number | boolean // (v-model) 当前选中的值
 }
 const props = withDefaults(defineProps<Props>(), {
   options: () => [],
   disabled: false,
   vertical: false,
-  value: null,
+  checked: false,
   gap: 8,
+  width: 'auto',
+  height: 'auto',
   button: false,
   buttonStyle: 'outline',
-  buttonSize: 'middle'
+  buttonSize: 'middle',
+  value: undefined
 })
-const emits = defineEmits(['update:value', 'change'])
-function onClick(value: any) {
-  if (value !== props.value) {
+const optionsAmount = computed(() => {
+  // 选项总数
+  return props.options.length
+})
+const maxWidth = computed(() => {
+  // 单选区域最大宽度
+  if (!props.button) {
+    if (typeof props.width === 'number') {
+      return props.width + 'px'
+    } else {
+      return props.width
+    }
+  } else {
+    return 'auto'
+  }
+})
+const maxHeight = computed(() => {
+  // 单选区域最大高度
+  if (!props.button) {
+    if (typeof props.height === 'number') {
+      return props.height + 'px'
+    } else {
+      return props.height
+    }
+  } else {
+    return 'auto'
+  }
+})
+const gapValue = computed(() => {
+  if (!props.button) {
+    if (!props.vertical && Array.isArray(props.gap)) {
+      return `${props.gap[1]}px ${props.gap[0]}px`
+    }
+    return `${props.gap}px`
+  } else {
+    return 0
+  }
+})
+const radioChecked = ref<boolean>()
+const optionsCheckedValue = ref<string | number | boolean>()
+const emits = defineEmits(['update:checked', 'update:value', 'change'])
+watchEffect(() => {
+  radioChecked.value = props.checked
+})
+watchEffect(() => {
+  optionsCheckedValue.value = props.value
+})
+function checkDisabled(disabled: boolean | undefined) {
+  if (disabled === undefined) {
+    return props.disabled
+  } else {
+    return disabled
+  }
+}
+function onClick(value: string | number | boolean) {
+  if (value !== optionsCheckedValue.value) {
+    optionsCheckedValue.value = value
     emits('update:value', value)
     emits('change', value)
+  }
+}
+function onChecked() {
+  if (!radioChecked.value) {
+    radioChecked.value = true
+    emits('update:checked', true)
+    emits('change', true)
   }
 }
 </script>
@@ -41,49 +109,81 @@ function onClick(value: any) {
       'radio-button-small': button && buttonSize === 'small',
       'radio-button-large': button && buttonSize === 'large'
     }"
-    :style="`gap: ${button ? 0 : gap}px;`"
+    :style="`--radio-gap: ${gapValue}; --radio-max-width: ${maxWidth}; --radio-max-height: ${maxHeight};`"
   >
-    <div
-      v-if="!button"
-      class="m-radio-wrap"
-      :class="{ 'radio-disabled': disabled || option.disabled }"
-      v-for="(option, index) in options"
-      :key="index"
-      @click="disabled || option.disabled ? () => false : onClick(option.value)"
-    >
-      <span class="radio-handle" :class="{ 'radio-checked': value === option.value }"></span>
-      <span class="radio-label">
-        <slot :label="option.label">{{ option.label }}</slot>
-      </span>
-    </div>
+    <template v-if="optionsAmount">
+      <template v-if="!button">
+        <div
+          class="m-radio-wrap"
+          :class="{ 'radio-disabled': checkDisabled(option.disabled) }"
+          v-for="(option, index) in options"
+          :key="index"
+          @click="checkDisabled(option.disabled) ? () => false : onClick(option.value)"
+        >
+          <span class="radio-handle" :class="{ 'radio-checked': optionsCheckedValue === option.value }"></span>
+          <span class="radio-label">
+            <slot :label="option.label">{{ option.label }}</slot>
+          </span>
+        </div>
+      </template>
+      <template v-else>
+        <div
+          tabindex="0"
+          class="m-radio-button-wrap"
+          :class="{
+            'radio-button-checked': optionsCheckedValue === option.value,
+            'radio-button-disabled': checkDisabled(option.disabled)
+          }"
+          v-for="(option, index) in options"
+          :key="index"
+          @click="checkDisabled(option.disabled) ? () => false : onClick(option.value)"
+        >
+          <span class="radio-label">
+            <slot :label="option.label">{{ option.label }}</slot>
+          </span>
+        </div>
+      </template>
+    </template>
     <template v-else>
-      <div
-        tabindex="0"
-        class="m-radio-button-wrap"
-        :class="{
-          'radio-button-checked': value === option.value,
-          'radio-button-disabled': disabled || option.disabled
-        }"
-        v-for="(option, index) in options"
-        :key="index"
-        @click="disabled || option.disabled ? () => false : onClick(option.value)"
-      >
-        <span class="radio-label">
-          <slot :label="option.label">{{ option.label }}</slot>
-        </span>
-      </div>
+      <template v-if="!button">
+        <div class="m-radio-wrap" :class="{ 'radio-disabled': disabled }" @click="disabled ? () => false : onChecked()">
+          <span class="radio-handle" :class="{ 'radio-checked': radioChecked }"></span>
+          <span class="radio-label">
+            <slot></slot>
+          </span>
+        </div>
+      </template>
+      <template v-else>
+        <div
+          tabindex="0"
+          class="m-radio-button-wrap"
+          :class="{
+            'radio-button-checked': radioChecked,
+            'radio-button-disabled': disabled
+          }"
+          @click="disabled ? () => false : onChecked()"
+        >
+          <span class="radio-label">
+            <slot></slot>
+          </span>
+        </div>
+      </template>
     </template>
   </div>
 </template>
 <style lang="less" scoped>
 .m-radio {
   display: inline-flex;
+  flex-wrap: wrap;
+  gap: var(--radio-gap);
   color: rgba(0, 0, 0, 0.88);
   font-size: 14px;
   line-height: 1;
+  max-width: var(--radio-max-width);
+  max-height: var(--radio-max-height);
+  overflow: auto;
   .m-radio-wrap {
-    height: 100%;
-    display: inline-flex; // 设置为flex布局后，所有的子元素都会变成行内块元素
+    display: inline-flex;
     align-items: flex-start;
     cursor: pointer;
     &:not(.radio-disabled):hover {
@@ -92,13 +192,13 @@ function onClick(value: any) {
       }
     }
     .radio-handle {
-      position: relative;
       /*
         如果所有项目的flex-shrink属性都为1，当空间不足时，都将等比例缩小
         如果一个项目的flex-shrink属性为0，其他项目都为1，则空间不足时，前者不缩小。
       */
       flex-shrink: 0; // 默认 1.即空间不足时，项目将缩小
-      margin-top: 3px;
+      position: relative;
+      top: 3px;
       width: 16px;
       height: 16px;
       background: transparent;
@@ -137,9 +237,7 @@ function onClick(value: any) {
     .radio-label {
       word-break: break-all;
       padding: 0 8px;
-      font-size: 14px;
-      line-height: 22px;
-      display: inline-block;
+      line-height: 1.5714285714285714;
     }
   }
   .radio-disabled {
@@ -218,8 +316,8 @@ function onClick(value: any) {
   }
 }
 .radio-vertical {
-  display: inline-flex;
   flex-direction: column;
+  flex-wrap: nowrap;
 }
 .radio-button-solid {
   .radio-button-checked:not(.radio-button-disabled) {